requirements.txt
# UI
gradio==4.44.1
gradio-client==1.3.0

# Parsing/HTML e imágenes
beautifulsoup4==4.12.3
lxml==5.2.2
pillow==10.4.0

# Server stack que Gradio usa por debajo
fastapi==0.112.0
starlette==0.37.2
pydantic==2.7.4
uvicorn==0.30.1

# Environment variables
python-dotenv==1.0.0